fix: correct io() fallback mechanism
If no Interactor has been provided in the first place, `io()` has a
fallback in which it creates a new Interactor instance from scratch.
Since the addition of php8 types in c94535f15e2d7617b98507876017a46bb99325dd,
this fails because of a missing default value for the io-property.

This adds a default `null` value to the property to make the property
check in `io()` work again.